+++ NEWS.md
@@ -1,5 +1,8 @@
 # holiday.dates 0.1.1
 * Added new function for Back to School: `Back2School()`
+* Added new functions for Target Circle Weeks: 
+    - `USTgt_Cirle_Wk_Summer()`
+    - `USTgt_Cirle_Wk_Fall()`
 
 # holiday.dates 0.1.0